fix(thumbnails): Refactoring upload logic, removing duplicated code, minor styling tweaks
