+++ src/components/EntityThumbnailUploader/EntityThumbnailUploader.styled.tsx
@@ -1,11 +1,6 @@
 import styled from "styled-components";
 
 export const DragAndDropWrapper = styled.div`
-  &.isCompact {
-    .icon {
-      font-size: 2rem;
-    }
-  }
 
   > input[type='file'] {
     display: none;
